+++ rh-ressource/typeRessource.php
@@ -33,7 +33,7 @@
 				//$ATMdb->db->debug=true;
 				$ressource->load($ATMdb, $_REQUEST['id']);
 				$ressource->set_values($_REQUEST);
-				$mesg = '<div class="ok">Modifications effectuées</div>';
+				$mesg = '<div class="ok">Modifications effectuées.</div>';
 				$mode = 'view';
 				if(isset($_REQUEST['TField'])){
 				
@@ -47,7 +47,7 @@
 					//ajout de ce champs à la classe ressource
 					$ressource->addField($ATMdb, $_REQUEST['TNField']);
 					
-					$mesg = '<div class="ok">Le champs a bien été créé</div>';
+					$mesg = '<div class="ok">Le champs a bien été créé.</div>';
 					$mode = 'edit';
 				}
 
@@ -62,31 +62,26 @@
 				_fiche($ATMdb, $ressource,'view');
 				break;
 		
-			case 'deleteField':
-				//$ATMdb->db->debug=true;
-				$ressource->delField($ATMdb, $_REQUEST['idField']);
-				$ressource->load($ATMdb, $_REQUEST['id']);
-				
-				$mesg = '<div class="ok">Le champs a bien été supprimé</div>';
-				$mode = 'edit';
-				_fiche($ATMdb, $ressource,$mode);
-				break;
-				
 			case 'delete':
 				$ressource->load($ATMdb, $_REQUEST['id']);
 				//$ATMdb->db->debug=true;
 				
 				//avant de supprimer, on vérifie qu'aucune ressource n'est de ce type. Sinon on ne le supprime pas.
-				if ( ! $ressource->isUsedByRessource($ATMdb)){
-					$ressource->delete($ATMdb);
-					?>
-					<script language="javascript">
-						document.location.href="?delete_ok=1";					
-					</script>
-					<?
+				if (!$ressource->isUsedByRessource($ATMdb)){
+					if ($ressource->delete($ATMdb)){
+						?>
+						<script language="javascript">
+							document.location.href="?delete_ok=1";					
+						</script>
+						<?	
+					}
+					else{
+						$mesg = '<div class="error">Ce type de ressource ne peut pas être supprimé.</div>';
+						_fiche($ATMdb, $ressource, 'view');
+					}
 				}
 				else{
-					$mesg = '<div class="ok">Le type de ressource est utilisé par une ressource. Il ne peut pas être supprimé.</div>';
+					$mesg = '<div class="error">Le type de ressource est utilisé par une ressource. Il ne peut pas être supprimé.</div>';
 					_fiche($ATMdb, $ressource, 'view');
 				}
